convert project from maven to gradle
