+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sat Mar 30 02:03:48 BRT 2019
+gradle.version=5.2.1
