+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2018 Iago Silva
+Copyright (c) 2018 Iago Paixão
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
