+++ build.gradle
@@ -1,0 +1,45 @@
+plugins {
+    id 'org.springframework.boot' version '2.2.0.M1'
+    id 'net.ltgt.apt' version '0.8'
+    id 'application'
+    id 'java'
+}
+
+apply plugin: 'io.spring.dependency-management'
+
+group = 'com.ipaixao'
+version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '11'
+targetCompatibility = '11'
+
+applicationName = 'car-graphql-api'
+
+repositories {
+    mavenCentral()
+    maven { url 'https://repo.spring.io/snapshot' }
+    maven { url 'https://repo.spring.io/milestone' }
+}
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-data-rest'
+    implementation 'org.springframework.boot:spring-boot-starter'
+
+    implementation 'com.graphql-java:graphql-spring-boot-starter:3.10.0'
+    implementation 'com.graphql-java:graphiql-spring-boot-starter:3.10.0'
+    implementation 'io.leangen.graphql:spqr:0.9.6'
+
+    compileOnly 'org.projectlombok:lombok:1.18.6'
+    annotationProcessor 'org.projectlombok:lombok:1.18.6'
+
+    implementation 'org.mapstruct:mapstruct:1.3.0.Final'
+    compileOnly 'org.mapstruct:mapstruct-processor:1.2.0.Final'
+
+    implementation 'org.apache.commons:commons-collections4:4.2'
+    runtime 'com.h2database:h2:1.4.196'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+}
+
+application {
+    mainClassName = 'com.ipaixao.cargraphqlapi.CarGraphqlApplication'
+}
