+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Mar 29 22:19:15 BRT 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.2.1-all.zip
