--- src/main/java/com/iagosilva/api/CarApplication.java
@@ -1,13 +1,0 @@
-package com.iagosilva.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CarApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CarApplication.class, args);
-	}
-}
-
