--- src/main/java/com/iagosilva/api/repository/CarRepository.java
@@ -1,8 +1,0 @@
-package com.iagosilva.api.repository;
-
-import com.iagosilva.api.domain.Car;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CarRepository extends JpaRepository<Car, Long> {}
