--- src/main/java/com/iagosilva/api/service/mapper/EntityMapper.java
@@ -1,14 +1,0 @@
-package com.iagosilva.api.service.mapper;
-
-import java.util.List;
-
-public interface EntityMapper<D, E> {
-
-    E toEntity(D dto);
-
-    D toDto(E entity);
-
-    List<E> toEntity(List<D> dtoList);
-
-    List<D> toDto(List<E> entityList);
-}
