--- src/main/java/com/iagosilva/api/util/GraphQLUtils.java
@@ -1,9 +1,0 @@
-package com.iagosilva.api.util;
-
-public final class GraphQLUtils {
-
-    private GraphQLUtils(){}
-
-    public static final String QUERY_PARAM = "query";
-    public static final String OPERATION_NAME = "operationName";
-}
