+++ src/main/java/com/ipaixao/cargraphqlapi/CarGraphqlApplication.java
@@ -1,0 +1,12 @@
+package com.ipaixao.cargraphqlapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CarGraphqlApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(CarGraphqlApplication.class, args);
+  }
+}
