+++ src/main/java/com/ipaixao/cargraphqlapi/repository/CarRepository.java
@@ -1,0 +1,8 @@
+package com.ipaixao.cargraphqlapi.repository;
+
+import com.ipaixao.cargraphqlapi.domain.Car;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CarRepository extends JpaRepository<Car, Long> {}
