+++ src/main/java/com/ipaixao/cargraphqlapi/service/CarService.java
@@ -1,0 +1,18 @@
+package com.ipaixao.cargraphqlapi.service;
+
+import com.ipaixao.cargraphqlapi.service.dto.CarDTO;
+
+import java.util.List;
+
+public interface CarService {
+
+  List<CarDTO> getCars();
+
+  CarDTO getCarById(Long id);
+
+  CarDTO saveCar(CarDTO car);
+
+  CarDTO updateCar(CarDTO car);
+
+  void deleteCar(Long id);
+}
