+++ src/main/java/com/ipaixao/cargraphqlapi/service/mapper/CarMapper.java
@@ -1,0 +1,15 @@
+package com.ipaixao.cargraphqlapi.service.mapper;
+
+import com.ipaixao.cargraphqlapi.domain.Car;
+import com.ipaixao.cargraphqlapi.service.dto.CarDTO;
+import org.mapstruct.*;
+
+import static org.mapstruct.NullValueCheckStrategy.ALWAYS;
+
+@Mapper(
+    componentModel = "spring",
+    injectionStrategy = InjectionStrategy.CONSTRUCTOR,
+    nullValueCheckStrategy = ALWAYS,
+    nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE
+)
+public interface CarMapper extends EntityMapper<CarDTO, Car> {}
