+++ src/main/java/com/ipaixao/cargraphqlapi/service/mapper/EntityMapper.java
@@ -1,0 +1,14 @@
+package com.ipaixao.cargraphqlapi.service.mapper;
+
+import java.util.List;
+
+public interface EntityMapper<D, E> {
+
+    E toEntity(D dto);
+
+    D toDto(E entity);
+
+    List<E> toEntity(List<D> dtoList);
+
+    List<D> toDto(List<E> entityList);
+}
