+++ src/main/java/com/ipaixao/cargraphqlapi/util/GraphQLUtils.java
@@ -1,0 +1,9 @@
+package com.ipaixao.cargraphqlapi.util;
+
+public final class GraphQLUtils {
+
+    private GraphQLUtils(){}
+
+    public static final String QUERY_PARAM = "query";
+    public static final String OPERATION_NAME = "operationName";
+}
