--- src/test/java/com/iagosilva/api/CarApplicationTests.java
@@ -1,14 +1,0 @@
-package com.iagosilva.api;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class CarApplicationTests {
-
-    @Test
-    public void contextLoads() {}
-}
