+++ src/test/java/com/ipaixao/cargraphqlapi/CarGraphqlApplicationTests.java
@@ -1,0 +1,14 @@
+package com.ipaixao.cargraphqlapi;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@SpringBootTest
+@RunWith(SpringRunner.class)
+public class CarGraphqlApplicationTests {
+
+  @Test
+  public void contextLoads() {}
+}
